Reconstructing the metrics folder and add type check logic
