--- src/otx/algo/classification/metrics/__init__.py
@@ -1,8 +1,0 @@
-# Copyright (C) 2023 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-"""Backbone modules for OTX custom model."""
-
-from .hlabel_accuracy import HLabelAccuracy
-
-__all__ = ["HLabelAccuracy"]
