+++ src/otx/core/model/module/action_classification.py
@@ -44,13 +44,17 @@ class OTXActionClsLitModule(OTXLitModule):
         )
 
         if metric:
-            sig = inspect.signature(metric)
-            param_dict = {}
-            for name, param in sig.parameters.items():
-                param_dict[name] = param.default if name != "num_classes" else self.model.num_classes
-            param_dict.pop("kwargs")
+            if inspect.isclass(metric):
+                sig = inspect.signature(metric)
+                param_dict = {}
+                for name, param in sig.parameters.items():
+                    param_dict[name] = param.default if name != "num_classes" else self.model.num_classes
+                param_dict.pop("kwargs", {})
+                metric = metric(**param_dict)
+            else:
+                msg = "Function based metric not yet supported."
+                raise ValueError(msg)
 
-            metric = metric(**param_dict)  # type: ignore[call-arg]
         self.metric = metric
 
     def _log_metrics(self, meter: Accuracy, key: str) -> None:
