+++ src/otx/core/model/module/action_detection.py
@@ -45,13 +45,17 @@ class OTXActionDetLitModule(OTXLitModule):
         )
 
         if metric:
-            sig = inspect.signature(metric)
-            param_dict = {}
-            for name, param in sig.parameters.items():
-                param_dict[name] = param.default
-            param_dict.pop("kwargs")
+            if inspect.isclass(metric):
+                sig = inspect.signature(metric)
+                param_dict = {}
+                for name, param in sig.parameters.items():
+                    param_dict[name] = param.default
+                param_dict.pop("kwargs")
+                metric = metric(**param_dict)
+            else:
+                msg = "Function based metric not yet supported."
+                raise ValueError(msg)
 
-            metric = metric(**param_dict)  # type: ignore[call-arg]
         self.metric = metric
 
     def _log_metrics(self, meter: MeanAveragePrecision, key: str) -> None:
