+++ src/otx/core/model/module/detection.py
@@ -45,13 +45,17 @@ class OTXDetectionLitModule(OTXLitModule):
             metric=metric,
         )
 
-        param_dict = {}
         if metric:
-            sig = inspect.signature(metric)
-            for name, param in sig.parameters.items():
-                param_dict[name] = param.default
-            param_dict.pop("kwargs")
-            metric = metric(**param_dict)  # type: ignore[call-arg]
+            if inspect.isclass(metric):
+                sig = inspect.signature(metric)
+                param_dict = {}
+                for name, param in sig.parameters.items():
+                    param_dict[name] = param.default
+                param_dict.pop("kwargs", {})
+                metric = metric(**param_dict)
+            else:
+                msg = "Function based metric not yet supported."
+                raise ValueError(msg)
 
         self.metric = metric
