+++ src/otx/core/model/module/segmentation.py
@@ -51,17 +51,21 @@ class OTXSegmentationLitModule(OTXLitModule):
             raise RuntimeError(msg)
 
         if metric:
-            sig = inspect.signature(metric)
-            param_dict = {}
-            for name, param in sig.parameters.items():
-                if name == "num_classes":
-                    param_dict[name] = self.model.num_classes + 1
-                    param_dict["ignore_index"] = self.model.num_classes
-                else:
-                    param_dict[name] = param.default
-            param_dict.pop("kwargs")
+            if inspect.isclass(metric):
+                sig = inspect.signature(metric)
+                param_dict = {}
+                for name, param in sig.parameters.items():
+                    if name == "num_classes":
+                        param_dict[name] = self.model.num_classes + 1
+                        param_dict["ignore_index"] = self.model.num_classes
+                    else:
+                        param_dict[name] = param.default
+                param_dict.pop("kwargs", {})
+                metric = metric(**param_dict)
+            else:
+                msg = "Function based metric not yet supported."
+                raise ValueError(msg)
 
-            metric = metric(**param_dict)  # type: ignore[call-arg]
         self.metric = metric
 
     def _log_metrics(self, meter: Dice, key: str) -> None:
