+++ src/otx/recipe/classification/h_label_cls/otx_deit_tiny.yaml
@@ -23,7 +23,7 @@ scheduler:
       monitor: val/accuracy
 
 metric:
-  class_path: otx.algo.classification.metrics.hlabel_accuracy.HLabelAccuracy
+  class_path: otx.core.metrics.HLabelAccuracy
   init_args:
     num_multiclass_heads: 0
     num_multilabel_classes: 0
